[spark] extract SparkCatalogBase
